chore: move switcher to components
